Commander improving

+++ inc/Enemie_forces.hpp
@@ -12,7 +12,7 @@ class Enemie_forces{
     public:
         int turn_number; //!< The turn number of the game.
         int base_stamina; //!< The base stamina of the enemy forces.
-        int number_of_units_relatively; //!< The number of enemy units relatively.
+        double number_of_units_relatively; //!< The number of enemy units relatively.
         int avr_stamina; //!< The average stamina of the enemy forces.
         int avr_speed; //!< The average speed of the enemy forces.
         int avr_attack_range; //!< The average attack range of the enemy forces.
